Refactoring some of the paths to make the code clearer, please note the use of @since in versioned files
